requirements.txt
ansible==2.9.2
argon2-cffi==20.1.0
asgiref==3.2.3
astroid==2.3.3
bcrypt==3.1.7
cachetools==4.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
Django==3.0.1
django-appconf==1.0.4
django-bootstrap3==12.1.0
django-mediumeditor==1.0.0
django-tinymce==3.0.2
Faker==4.1.0
google-api-core==1.17.0
google-api-python-client==1.8.4
google-auth==1.16.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
httplib2==0.18.1
idna==2.8
isort==4.3.21
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
misaka==2.1.1
oauthlib==3.1.0
Pillow==7.1.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
typed-ast==1.4.0
uritemplate==3.0.1
urllib3==1.25.7
virtualenv==16.7.9
wrapt==1.11.2
